feat(layout): use layoutin products page
