+++ components/Header/index.tsx
@@ -4,15 +4,17 @@ import { theme, Layout, Button } from 'antd';
 const { Header } = Layout;
 
 const GeneralHeader = (props) => {
-    const { handleShowCreateItemModal, text } = props;
+    const { handleShowCreateItemModal, ctaText, hasCta } = props;
     const {
         token: { colorBgContainer },
     } = theme.useToken();
     return (
-        <Header style={{ padding: 0, background: colorBgContainer }}>
-            <Button size="large" type="primary" shape="round" onClick={handleShowCreateItemModal}>
-                {text}
-            </Button>
+        <Header style={{ padding: 3, background: colorBgContainer }}>
+            {hasCta && (
+                <Button size="large" type="primary" shape="round" onClick={handleShowCreateItemModal}>
+                    {ctaText}
+                </Button>
+            )}
         </Header>
     )
 };
