+++ components/IconText/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Space } from 'antd';
+
+const IconText = ({ icon, text }: { icon: React.FC; text: string }) => (
+    <Space>
+        {React.createElement(icon)}
+        {text}
+    </Space>
+);
+
+export default IconText
