+++ pages/dashboard.tsx
@@ -43,7 +43,7 @@ const Dashboard = () => {
                 handleCancel={handleCancel}
                 confirmLoading={confirmLoading}
             />
-            <GeneralLayout handleShowCreateItemModal={handleShowCreateItemModal}>
+            <GeneralLayout handleShowCreateItemModal={handleShowCreateItemModal} hasCta ctaText="Create Order">
                 <Row gutter={16}>
                     <Col span={12}>
                         <Card bordered={false}>
