+++ pages/products/index.tsx
@@ -1,130 +1,11 @@
-import React, { useState } from 'react';
-import {
-    UserOutlined,
-    BellOutlined,
-    LikeOutlined,
-    StarOutlined,
-    MessageOutlined,
-    AppstoreOutlined,
-    CodeSandboxOutlined,
-    NodeCollapseOutlined,
-    DeploymentUnitOutlined,
-} from '@ant-design/icons';
-import {
-    Menu,
-    List,
-    theme,
-    Space,
-    Steps,
-    Modal,
-    Layout,
-    Button,
-    Avatar,
-    message,
-} from 'antd';
+import React from 'react';
+import { List, Avatar } from 'antd';
+import { LikeOutlined, StarOutlined, MessageOutlined } from '@ant-design/icons';
 
-const { Header, Content, Footer, Sider } = Layout;
+import IconText from '../../components/IconText';
+import GeneralLayout from '../../components/Layout/General';
 
-const siderOptions = [
-    {
-        icon: <AppstoreOutlined />,
-        label: 'Dashboard'
-    },
-    {
-        icon: <CodeSandboxOutlined />,
-        label: 'Products'
-    },
-    {
-        icon: <DeploymentUnitOutlined />,
-        label: 'Shippments'
-    },
-    {
-        icon: <BellOutlined />,
-        label: 'Notifications'
-    },
-    {
-        icon: <NodeCollapseOutlined />,
-        label: 'Returns'
-    },
-    {
-        icon: <UserOutlined />,
-        label: 'Profile'
-    }
-];
-const IconText = ({ icon, text }: { icon: React.FC; text: string }) => (
-    <Space>
-        {React.createElement(icon)}
-        {text}
-    </Space>
-);
-const items = siderOptions.map((entry, index) => {
-    const key = index + 1;
-    return {
-        key,
-        icon: entry.icon,
-        label: entry.label,
-    };
-});
 const Products = () => {
-    const [open, setOpen] = useState(false);
-    const [current, setCurrent] = useState(0);
-    const [modalText, setModalText] = useState('Content of the modal');
-    const [confirmLoading, setConfirmLoading] = useState(false);
-    const {
-        token: { colorBgContainer },
-    } = theme.useToken();
-    const handleShowCreateItemModal = () => {
-        setOpen(true);
-    };
-    const handleOk = () => {
-        setModalText('The modal will be closed after two seconds');
-        setConfirmLoading(true);
-        setTimeout(() => {
-            setOpen(false);
-            setConfirmLoading(false);
-        }, 2000);
-    };
-    const handleCancel = () => {
-        console.log('Clicked cancel button');
-        setOpen(false);
-    };
-    const normFile = (e: any) => {
-        if (Array.isArray(e)) {
-            return e;
-        }
-        return e?.fileList;
-    };
-    const steps = [
-        {
-            title: 'First',
-            content: 'First-content',
-        },
-        {
-            title: 'Second',
-            content: 'Second-content',
-        },
-        {
-            title: 'Last',
-            content: 'Last-content',
-        },
-    ];
-    const next = () => {
-        setCurrent(current + 1);
-    };
-    const prev = () => {
-        setCurrent(current - 1);
-    };
-    const { token } = theme.useToken();
-    const itemsss = steps.map((item) => ({ key: item.title, title: item.title }));
-    const contentStyle: React.CSSProperties = {
-        lineHeight: '260px',
-        textAlign: 'center',
-        color: token.colorTextTertiary,
-        backgroundColor: token.colorFillAlter,
-        borderRadius: token.borderRadiusLG,
-        border: `1px dashed ${token.colorBorder}`,
-        marginTop: 16,
-    };
     const data = Array.from({ length: 23 }).map((_, i) => ({
         href: 'https://ant.design',
         title: `ant design part ${i}`,
@@ -135,100 +16,48 @@ const Products = () => {
             'We supply a series of design principles, practical patterns and high quality design resources (Sketch and Axure), to help people create their product prototypes beautifully and efficiently.',
     }));
     return (
-        <Layout hasSider>
-            {open && (
-                <Modal
-                    title="Title"
-                    open={open}
-                    onOk={handleOk}
-                    confirmLoading={confirmLoading}
-                    onCancel={handleCancel}
-                >
-                    <Steps current={current} items={itemsss} />
-                    <div style={contentStyle}>{steps[current].content}</div>
-                    <div style={{ marginTop: 24 }}>
-                        {current < steps.length - 1 && (
-                            <Button type="primary" onClick={() => next()}>
-                                Next
-                            </Button>
-                        )}
-                        {current === steps.length - 1 && (
-                            <Button type="primary" onClick={() => message.success('Processing complete!')}>
-                                Done
-                            </Button>
-                        )}
-                        {current > 0 && (
-                            <Button style={{ margin: '0 8px' }} onClick={() => prev()}>
-                                Previous
-                            </Button>
-                        )}
-                    </div>
-                </Modal>
-            )}
-            <Sider
-                style={{
-                    overflow: 'auto',
-                    height: '100vh',
-                    position: 'fixed',
-                    left: 0,
-                    top: 0,
-                    bottom: 0,
+        <GeneralLayout handleShowCreateItemModal={() => { }} hasCta={false}>
+            <List
+                itemLayout="vertical"
+                size="large"
+                pagination={{
+                    onChange: (page) => {
+                        console.log(page);
+                    },
+                    pageSize: 3,
                 }}
-            >
-                <div className="demo-logo-vertical" />
-                <Menu theme="dark" mode="inline" defaultSelectedKeys={['1']} items={items} />
-            </Sider>
-            <Layout className="site-layout" style={{ marginLeft: 200 }}>
-                <Header style={{ padding: 0, background: colorBgContainer }}>
-                    <Button size="large" type="primary" shape="round" onClick={handleShowCreateItemModal}>
-                        Create Item
-                    </Button>
-                </Header>
-                <Content style={{ margin: '24px 16px 0', overflow: 'initial' }}>
-                    <List
-                        itemLayout="vertical"
-                        size="large"
-                        pagination={{
-                            onChange: (page) => {
-                                console.log(page);
-                            },
-                            pageSize: 3,
-                        }}
-                        dataSource={data}
-                        footer={
-                            <div>
-                                <b>ant design</b> footer part
-                            </div>
+                dataSource={data}
+                footer={
+                    <div>
+                        <b>ant design</b> footer part
+                    </div>
+                }
+                renderItem={(item) => (
+                    <List.Item
+                        key={item.title}
+                        actions={[
+                            <IconText icon={StarOutlined} text="156" key="list-vertical-star-o" />,
+                            <IconText icon={LikeOutlined} text="156" key="list-vertical-like-o" />,
+                            <IconText icon={MessageOutlined} text="2" key="list-vertical-message" />,
+                        ]}
+                        extra={
+                            <img
+                                width={272}
+                                alt="logo"
+                                src="https://gw.alipayobjects.com/zos/rmsportal/mqaQswcyDLcXyDKnZfES.png"
+                            />
                         }
-                        renderItem={(item) => (
-                            <List.Item
-                                key={item.title}
-                                actions={[
-                                    <IconText icon={StarOutlined} text="156" key="list-vertical-star-o" />,
-                                    <IconText icon={LikeOutlined} text="156" key="list-vertical-like-o" />,
-                                    <IconText icon={MessageOutlined} text="2" key="list-vertical-message" />,
-                                ]}
-                                extra={
-                                    <img
-                                        width={272}
-                                        alt="logo"
-                                        src="https://gw.alipayobjects.com/zos/rmsportal/mqaQswcyDLcXyDKnZfES.png"
-                                    />
-                                }
-                            >
-                                <List.Item.Meta
-                                    avatar={<Avatar src={item.avatar} />}
-                                    title={<a href={item.href}>{item.title}</a>}
-                                    description={item.description}
-                                />
-                                {item.content}
-                            </List.Item>
-                        )}
-                    />
-                </Content>
-                <Footer style={{ textAlign: 'center' }}>IOTA - Logistics</Footer>
-            </Layout>
-        </Layout>
+                    >
+                        <List.Item.Meta
+                            avatar={<Avatar src={item.avatar} />}
+                            title={<a href={item.href}>{item.title}</a>}
+                            description={item.description}
+                        />
+                        {item.content}
+                    </List.Item>
+                )}
+            />
+        </GeneralLayout>
     );
 };
